fix(rpcsmartrouter): chain the health arms, quiet the session-cap log
Two defects found in review, both cheaper to fix than to argue for leaving.

The endpoint-health arms in relayInnerDirect were two independent ifs, so one answer could
blame the endpoint and then certify it. On JSON-RPC and gRPC that cannot arise — a body error
always sets IsNodeError and the reset gate reads it — but the two gates read different fields,
and on REST IsNodeError comes from the HTTP status alone. A 200 carrying a fault body would
count the endpoint up and immediately reset it to zero, wiping the whole failure history. No
registry row produces that shape today; chaining the arms is what keeps it impossible when a
REST matcher for a node-fault code is added, which this branch's own comments invite.

The session-cap log fired at ERROR once per REQUEST. Dropping the block for
MaximumNumberOfBlockListedSessionsError leaves the address in validAddresses, so every
subsequent request re-selects it, fails to get a session, and logs again for the rest of the
epoch — where the block previously made it fire once. It also logged the whole validAddresses
slice, read without csm.lock, which the surrounding code avoids for exactly that reason and
which this change makes far more likely to be hit. Capacity is not an error: the two cap cases
now log at debug without the slice, and the ERROR line is kept for the branch that is genuinely
unexpected.

Also drops a uint64() conversion that became redundant when the threshold stopped being an int.

+++ protocol/lavasession/consumer_types.go
@@ -482,7 +482,7 @@ func (e *Endpoint) markUnhealthyAt(at time.Time) {
 		utils.LavaFormatDebug("endpoint approaching the disable threshold",
 			utils.LogAttr("endpoint", addr),
 			utils.LogAttr("refusals", refusals),
-			utils.LogAttr("threshold", uint64(MaxConsecutiveConnectionAttempts)),
+			utils.LogAttr("threshold", MaxConsecutiveConnectionAttempts),
 			utils.LogAttr("remaining", MaxConsecutiveConnectionAttempts-refusals),
 			utils.LogAttr("is_direct_rpc", isDirect),
 		)
